+++ src/alley/wp/match-blocks.php
@@ -20,9 +20,11 @@ use Laminas\Validator\ValidatorChain;
 /**
  * Match blocks within the given content.
  *
- * @param array|int|\WP_Post|string $source Array of blocks, post ID or object with blocks in `post_content`, or string of block HTML.
- * @param array                     $args   {
- *    Optional. Array of arguments for matching which blocks to return. The defaults serve to match all non-empty blocks.
+ * @param int|\WP_Post|string|array[]|\WP_Block_Parser_Block|array $source Post ID or object with blocks in `post_content`, string of block HTML,
+ *                                                                         array of blocks, or a single block instance. Passing a single block
+ *                                                                         will return matches from its inner blocks.
+ * @param array                                                    $args   {
+ *    Optional. Array of arguments for matching which blocks to return. The defaults match all non-empty blocks.
  *
  *    @type array                     $attrs             {
  *        Match blocks with the given attributes.
@@ -81,6 +83,10 @@ function match_blocks( $source, $args = [] ) {
 	$blocks = [];
 	$error  = $args['count'] ? 0 : [];
 
+	if ( $source instanceof \WP_Block_Parser_Block ) {
+		$source = (array) $source;
+	}
+
 	if ( \is_array( $source ) ) {
 		$blocks = $source;
 	}
@@ -99,7 +105,11 @@ function match_blocks( $source, $args = [] ) {
 		$blocks = parse_blocks( $source );
 	}
 
-	if ( ! \is_array( $blocks ) || 0 === \count( $blocks ) ) {
+	if ( \is_array( $blocks ) && isset( $blocks['innerBlocks'] ) ) {
+		$blocks = $blocks['innerBlocks'];
+	}
+
+	if ( ! wp_is_numeric_array( $blocks ) || 0 === \count( $blocks ) ) {
 		return $error;
 	}